Translations 3. (toasts)

+++ src/screens/ride/PassengerScreen.tsx
@@ -8,9 +8,10 @@ import { InfoTimeline, InfoTimelineItem } from "../../components/info/InfoTimeli
 import { ScreenView } from "../../components/screenView/ScreenView.tsx";
 import { PickerItemType } from "../../components/Input/picker/PickerItem.tsx";
 import { router } from "expo-router";
-import { DeleteExpenseToast } from "../../features/expense/presets/toasts/DeleteExpenseToast.ts";
 import { useAlert } from "../../ui/alert/hooks/useAlert.ts";
 import { useTranslation } from "react-i18next";
+import { DeleteToast, NotFoundToast } from "../../ui/alert/presets/toast/index.ts";
+import { DeleteModal } from "../../ui/alert/presets/modal/index.ts";
 
 export function PassengerScreen() {
     const { t } = useTranslation();
@@ -48,7 +49,7 @@ export function PassengerScreen() {
     const openCreateForm = () => router.push("/ride/passenger/create");
 
     const onEdit = (id: string, callback?: () => void) => {
-        if(!id) return;
+        if(!id) return openToast(NotFoundToast.warning(t("passengers.title_singular")));
 
         callback?.();
         router.push({
@@ -58,10 +59,12 @@ export function PassengerScreen() {
     };
 
     const handleDelete = useCallback(async (id: string, callback?: () => void) => {
+        if(!id) return openToast(NotFoundToast.warning(t("passengers.title_singular")));
+
         try {
             await passengerDao.delete(id);
 
-            openToast(DeleteExpenseToast.success());
+            openToast(DeleteToast.success(t("passengers.title_singular")));
             callback?.();
             await refresh();
             requestAnimationFrame(() => {
@@ -69,17 +72,15 @@ export function PassengerScreen() {
             });
         } catch(e) {
             console.log(e);
-            openToast(DeleteExpenseToast.error());
+            openToast(DeleteToast.error(t("passengers.title_singular")));
         }
     }, [passengerDao]);
 
-    const onDelete = useCallback((id: string) => {
-        openModal({
-            title: t("passengers.modal.delete"),
-            body: t("modal.delete_message"),
-            acceptText: t("form_button.delete"),
-            acceptAction: () => handleDelete(id)
-        });
+    const onDelete = useCallback((id: string, callback?: () => void) => {
+        openModal(DeleteModal({
+            name: t("passengers.title_singular"),
+            acceptAction: () => handleDelete(id, callback)
+        }));
     }, [openToast, openModal]);
 
     return (
